IGNITE-19975 Introduced abstract SSL Context Factory.
